Refactor : semantic-analyzer norm
