+++ src/manda/Makefile
@@ -7,6 +7,7 @@ endif
 
 NAME := minishell
 SRCS := main_syntax_analyzer_print.c \
+		print.c \
 
 # SRCS(in SRCDIR) ------
 DIRS := insert_data \
